Remove code comments
